requirements/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fairseq==0.12.2
ipykernel==6.29.5
joblib==1.4.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
praat-parselmouth==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
soundfile==0.12.1
tensorboard==2.17.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
transformers==4.34.0
